+++ apps/backoffice-v2/src/lib/blocks/components/ChildDocumentBlocks/ChildDocumentBlocks.tsx
@@ -1,11 +1,9 @@
-import { ctw } from '@/common/utils/ctw/ctw';
-import { CardContent } from '@/common/components/atoms/Card/Card.Content';
-import { Card } from '@/common/components/atoms/Card/Card';
 import { FunctionComponent } from 'react';
 import { TWorkflowById } from '@/domains/workflows/fetchers';
 import { UnknownRecord } from '@/common/types';
-import { cells } from '@/lib/blocks/cells';
 import { useChildDocumentBlocksLogic } from '@/lib/blocks/components/ChildDocumentBlocks/hooks/useChildDocumentBlocksLogic/useChildDocumentBlocksLogic';
+import { cells } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
+import { BlocksComponent } from '@ballerine/blocks';
 
 interface IChildDocumentBlocksProps {
   parentWorkflowId: string;
@@ -38,33 +36,9 @@ export const ChildDocumentBlocks: FunctionComponent<IChildDocumentBlocksProps> =
     isLoadingReuploadNeeded,
   });
 
-  if (!Array.isArray(childDocumentBlocks) || !childDocumentBlocks?.length) return null;
-
   return (
-    <>
-      {childDocumentBlocks?.map((childDocumentBlock, index) => (
-        <Card
-          key={index}
-          className={ctw(
-            'me-4 shadow-[0_4px_4px_0_rgba(174,174,174,0.0625)]',
-            childDocumentBlock.className,
-          )}
-        >
-          <CardContent
-            className={ctw('grid gap-2', {
-              'grid-cols-2': childDocumentBlock?.cells.some(
-                cell => cell?.type === 'multiDocuments',
-              ),
-            })}
-          >
-            {childDocumentBlock?.cells.map((cell, index) => {
-              const Cell = cells[cell?.type];
-
-              return <Cell key={index} {...cell} />;
-            })}
-          </CardContent>
-        </Card>
-      ))}
-    </>
+    <BlocksComponent blocks={childDocumentBlocks} cells={cells}>
+      {(Cell, cell) => <Cell {...cell} />}
+    </BlocksComponent>
   );
 };