+++ apps/backoffice-v2/src/lib/blocks/components/Container/interfaces.ts
@@ -1,5 +1,5 @@
 import { ComponentProps } from 'react';
-import { cells } from '@/lib/blocks/cells';
+import { cells } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
 
 export interface IContainerProps {
   id: string;