+++ apps/backoffice-v2/src/lib/blocks/create-blocks-typed/create-blocks-typed.ts
@@ -1,22 +1,35 @@
-import { BadgeCell } from '@/lib/blocks/components/BadgeCell/BadgeCell';
-import { CallToActionLegacy } from '@/lib/blocks/components/CallToActionLegacy/CallToActionLegacy';
-import { CallToAction } from '@/lib/blocks/components/CallToAction/CallToAction';
-import { DirectorsCallToAction } from '@/lib/blocks/components/DirectorsCallToAction';
+import { CellsMap, createBlocks } from '@ballerine/blocks';
+import { BlockCell } from '@/lib/blocks/components/BlockCell/BlockCell';
+import { TCell } from '@/lib/blocks/create-blocks-typed/types';
 import { FaceComparison } from '@/lib/blocks/components/FaceComparison/FaceComparison';
 import { Details } from '@/lib/blocks/components/Details/Details';
+import { NestedDetails } from '@/lib/blocks/components/NestedDetails/NestedDetails';
 import { MultiDocuments } from '@/lib/blocks/components/MultiDocuments/MultiDocuments';
 import { MapCell } from '@/lib/blocks/components/MapCell/MapCell';
 import { CaseCallToActionLegacy } from '@/lib/blocks/components/CaseCallToActionLegacy/CaseCallToActionLegacy';
 import { Paragraph } from '@/lib/blocks/components/Paragraph/Paragraph';
 import { DialogCell } from '@/lib/blocks/components/DialogCell/DialogCell';
-import { NestedDetails } from '@/lib/blocks/components/NestedDetails/NestedDetails';
-import { Container } from '@/lib/blocks/components/Container/Container';
-import { Alert } from '@/lib/blocks/components/Alert/Alert';
 import { Heading } from '@/lib/blocks/components/Heading/Heading';
 import { Subheading } from '@/lib/blocks/components/Subheading/Subheading';
+import { Alert } from '@/common/components/atoms/Alert/Alert';
+import { BadgeCell } from '@/lib/blocks/components/BadgeCell/BadgeCell';
+import { Container } from '@/lib/blocks/components/Container/Container';
+import { CallToActionLegacy } from '@/lib/blocks/components/CallToActionLegacy/CallToActionLegacy';
+import { CallToAction } from '@/lib/blocks/components/CallToAction/CallToAction';
+import { DirectorsCallToAction } from '@/lib/blocks/components/DirectorsCallToAction';
 import { TableCell } from '@/lib/blocks/components/TableCell/TableCell';
 
-export const cells = {
+export const createBlocksTyped = () => createBlocks<TCell>();
+
+const blocks = createBlocksTyped();
+
+declare module '@ballerine/blocks' {
+  interface BlocksClient {
+    cells: typeof blocks;
+  }
+}
+
+export const cells: CellsMap = {
   heading: Heading,
   subheading: Subheading,
   alert: Alert,
@@ -34,4 +47,5 @@ export const cells = {
   table: TableCell,
   paragraph: Paragraph,
   dialog: DialogCell,
+  block: BlockCell,
 };