+++ apps/backoffice-v2/src/lib/blocks/motion-badge-props.ts
@@ -1,7 +1,7 @@
 import { ComponentProps } from 'react';
 import { MotionBadge } from '../../common/components/molecules/MotionBadge/MotionBadge';
 
-export const motionProps: ComponentProps<typeof MotionBadge> = {
+export const motionBadgeProps: ComponentProps<typeof MotionBadge> = {
   exit: { opacity: 0, transition: { duration: 0.2 } },
   initial: { y: 10, opacity: 0 },
   transition: { type: 'spring', bounce: 0.3 },