+++ apps/backoffice-v2/src/lib/blocks/variants/KybDemoBlocks/KybDemoBlocks.tsx
@@ -1,7 +1,8 @@
-import { BlocksComponent } from '@/lib/blocks/components/BlocksComponent/BlocksComponent';
+import { BlocksComponent } from '@ballerine/blocks';
 import { useKybDemoBlocksLogic } from '@/lib/blocks/variants/KybDemoBlocks/hooks/useKybDemoBlocksLogic/useKybDemoBlocksLogic';
 import { NoBlocks } from '@/lib/blocks/components/NoBlocks/NoBlocks';
 import { ChildDocumentBlocks } from '@/lib/blocks/components/ChildDocumentBlocks/ChildDocumentBlocks';
+import { cells } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
 
 export const KybDemoBlocks = () => {
   const {
@@ -16,17 +17,21 @@ export const KybDemoBlocks = () => {
 
   return (
     <>
-      <BlocksComponent blocks={blocks} />
-      {kybChildWorkflows?.map(childWorkflow => (
-        <ChildDocumentBlocks
-          parentWorkflowId={workflowId}
-          childWorkflow={childWorkflow}
-          parentMachine={parentMachine}
-          key={childWorkflow?.id}
-          onReuploadNeeded={onReuploadNeeded}
-          isLoadingReuploadNeeded={isLoadingReuploadNeeded}
-        />
-      ))}
+      <BlocksComponent blocks={blocks} cells={cells}>
+        {(Cell, cell) => <Cell {...cell} />}
+      </BlocksComponent>
+      {Array.isArray(kybChildWorkflows) &&
+        !!kybChildWorkflows?.length &&
+        kybChildWorkflows?.map(childWorkflow => (
+          <ChildDocumentBlocks
+            parentWorkflowId={workflowId}
+            childWorkflow={childWorkflow}
+            parentMachine={parentMachine}
+            key={childWorkflow?.id}
+            onReuploadNeeded={onReuploadNeeded}
+            isLoadingReuploadNeeded={isLoadingReuploadNeeded}
+          />
+        ))}
       {!isLoading && !blocks?.length && !kybChildWorkflows?.length && <NoBlocks />}
     </>
   );