+++ apps/backoffice-v2/src/lib/blocks/variants/ManualReviewBlocks/ManualReviewBlocks.tsx
@@ -1,13 +1,16 @@
-import { BlocksComponent } from '@/lib/blocks/components/BlocksComponent/BlocksComponent';
+import { BlocksComponent } from '@ballerine/blocks';
 import { useManualReviewBlocksLogic } from '@/lib/blocks/variants/ManualReviewBlocks/hooks/useManualReviewBlocksLogic/useManualReviewBlocksLogic';
 import { NoBlocks } from '@/lib/blocks/components/NoBlocks/NoBlocks';
+import { cells } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
 
 export const ManualReviewBlocks = () => {
   const { blocks, isLoading } = useManualReviewBlocksLogic();
 
   return (
     <>
-      <BlocksComponent blocks={blocks} />
+      <BlocksComponent blocks={blocks} cells={cells}>
+        {(Cell, cell) => <Cell {...cell} />}
+      </BlocksComponent>
       {!isLoading && !blocks?.length && <NoBlocks />}
     </>
   );